requirements.txt
Flask~=3.1.0
Flask
Flask-Cors
pandas
python-dotenv
chromadb
google-generativeai
